fix(apps): satisfy multi-gateway checks
